# apps/ade-api/src/ade_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Automatic Data Extractor API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/workspaces": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create a new workspace",
        "operationId": "create_workspace_api_v1_workspaces_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to manage workspaces."
          },
          "403": {
            "description": "Administrator role required to create workspaces."
          },
          "404": {
            "description": "Specified owner could not be found or is inactive."
          },
          "409": {
            "description": "Workspace slug already exists."
          },
          "422": {
            "description": "Workspace name or slug is invalid."
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List workspaces for the authenticated user",
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "include_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Total"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePage"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to list workspaces."
          },
          "403": {
            "description": "Service account credentials cannot access workspace listings."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Retrieve workspace context by identifier",
        "operationId": "read_workspace_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to view workspace context."
          },
          "403": {
            "description": "Workspace access denied for the authenticated user."
          },
          "404": {
            "description": "Workspace not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "workspaces"
        ],
        "summary": "Update workspace metadata",
        "operationId": "update_workspace_api_v1_workspaces__workspace_id__patch",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to update workspaces."
          },
          "403": {
            "description": "Workspace permissions do not allow settings management."
          },
          "404": {
            "description": "Workspace not found."
          },
          "409": {
            "description": "Workspace slug already exists."
          },
          "422": {
            "description": "Workspace name or slug is invalid."
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "delete": {
        "tags": [
          "workspaces"
        ],
        "summary": "Delete a workspace",
        "operationId": "delete_workspace_api_v1_workspaces__workspace_id__delete",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required to delete workspaces."
          },
          "403": {
            "description": "Workspace permissions do not allow workspace deletion."
          },
          "404": {
            "description": "Workspace not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/default": {
      "put": {
        "tags": [
          "workspaces"
        ],
        "summary": "Mark a workspace as the caller's default",
        "operationId": "set_default_workspace_api_v1_workspaces__workspace_id__default_put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Workspace identifier",
              "title": "Workspace Id"
            },
            "description": "Workspace identifier"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required to set the default workspace."
          },
          "403": {
            "description": "Workspace access denied for the authenticated user."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "description": "UUIDv7 (RFC 9562) generated in the application layer."
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "WorkspaceCreate",
        "description": "Payload for creating a workspace."
      },
...
      "WorkspaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "UUIDv7 (RFC 9562) generated in the application layer."
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "roles",
          "permissions",
          "is_default"
        ],
        "title": "WorkspaceOut",
        "description": "Workspace information decorated with membership metadata."
      },
      "WorkspacePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "has_next",
          "has_previous"
        ],
        "title": "WorkspacePage",
        "description": "Paginated workspace listing."
      },
      "WorkspaceUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WorkspaceUpdate",
        "description": "Payload for updating workspace metadata."
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ade_session",
        "description": "Browser session cookie issued after interactive sign-in."
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer access token returned by ADE or an identity provider."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key for service integrations."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "security": [
    {
      "SessionCookie": []
    },
    {
      "HTTPBearer": []
    },
    {
      "APIKeyHeader": []
    }
  ]
}